move pagination to pkg, add sorting before paginated response, align the datatype
