+++ internal/store/postgres/organization_repository.go
@@ -210,16 +210,18 @@ func (r OrganizationRepository) List(ctx context.Context, flt organization.Filte
 		return []organization.Organization{}, fmt.Errorf("%w: %s", queryErr, err)
 	}
 
-	var totalCount uint
+	var totalCount int32
 	if err = r.dbc.WithTimeout(ctx, TABLE_ORGANIZATIONS, "Count", func(ctx context.Context) error {
 		return r.dbc.GetContext(ctx, &totalCount, totalCountQuery)
 	}); err != nil {
 		return nil, fmt.Errorf("%w: %s", dbErr, err)
 	}
 
-	flt.Pagination.SetTotalPages(totalCount)
+	flt.Pagination.SetCount(totalCount)
 
-	query, params, err := stmt.Limit(limit).Offset(offset).ToSQL()
+	stmt = stmt.Limit(uint(limit)).Offset(uint(offset)).Order(goqu.C("created_at").Desc())
+
+	query, params, err := stmt.ToSQL()
 	if err != nil {
 		return []organization.Organization{}, fmt.Errorf("%w: %s", queryErr, err)
 	}
